# Optimize combination_with_replacement_index