### Issue reference

Closes #772

### Changes

- Optimize **combination_with_replacement_index** by using an accumulating loop variable instead of summing up all elements up to index k per loop iteration

### Benchmarks

I used the following in IPython as a rudimentary benchmark.

**OS:** Linux x64 **CPU:** AMD Ryzen 7 5800X

```python
from more_itertools import combination_with_replacement_index
%timeit combination_with_replacement_index('adf', 'abcdefg')
# Before: 1.96 µs ± 9.67 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)
# After: 1.56 µs ± 3.66 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)
```